code/src/backend/requirements.txt
# Core Dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic>=2.4.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart==0.0.9

# OpenAI Integration
openai==1.12.0

# Data Processing
numpy==1.26.3
pandas==2.2.0
scikit-learn>=1.3.2
openpyxl==3.1.2
transformers==4.35.2
torch==2.1.1

# Additional Dependencies
jinja2==3.1.3
python-magic==0.4.27
aiofiles==23.2.1 